server/backend/requirements.txt
flask==1.1.2
flask-RESTful==0.3.9
Werkzeug==1.0.1
gunicorn==20.1.0
flask-cors==3.0.10
flask-migrate==2.1.1
flask-sqlalchemy==2.5.1
flask_script==2.0.6
psycopg2==2.8.6
sqlalchemy==1.4.15
python-dotenv==0.17.1
PyYAML==5.4.1
numpy>=1.17.4
pandas>=0.25.3
